Add login + registration

+++ routes/index.js
@@ -1,6 +1,10 @@
+const nameOfSort = ['id asc', 'id desc', 'value asc', 'value desc'];
+
 let express = require('express');
 let router = express.Router();
 let mysql      = require('mysql');
+let passport = require('passport');
+
 let connection = mysql.createConnection({
 	host     : 'localhost',
 	user     : 'root',
@@ -11,20 +15,20 @@ connection.connect(function(err) {
 	if (err) throw err
 });
 
-router.get('/', function(req, res, next) {
-	let field = 'ID', order = 'ASC';
 
-	if (req.query.sort) {
-		let sort = req.query.sort;
-		field = (sort.indexOf('ID') == -1) ? 'value' : 'ID';
-		order = (sort.indexOf('asc') != -1) ? 'ASC' : 'DESC';
+router.get('/', function(req, res, next) {
+	
+	if (!req.user) res.render('startPage', { rows: [], user : req.user });
+	
+	let query = "SELECT * FROM ALT_Database.express WHERE user_id=" + req.user.ID;
+	
+	if (req.query.sort && nameOfSort.includes(req.query.sort)) {
+		query += (" ORDER BY " + req.query.sort);
 	}
-	let query = 'SELECT ID, value FROM ALT_Database.express ORDER BY ' + 
-	field + ' ' + order;
 	
 	connection.query(query, function(err, rows, fields) {
 		if (err) return console.log(err);
-		res.render('startPage', { rows });
+		res.render('startPage', { rows, user : req.user });
 	});
 
 	
@@ -36,7 +40,7 @@ router.post('/', function(req, res, next) {
 		return res.status(500).json({ ID: 'None', Value: 'None', error: 'Server error.' });
 	}
 	
-	let query = "INSERT INTO ALT_Database.express (`value`) VALUES (\'" + req.body.value + "\')";
+	let query = "INSERT INTO ALT_Database.express (`value`,`user_id`) VALUES ('" + req.body.value + "'," + req.user.ID + ")";
 	
 	connection.query(query, function(err, result) {
 		if (err) { 
@@ -54,10 +58,7 @@ router.post('/:id', function(req, res, next) {
 		return res.status(500).json({ ID: 'None', Value: 'None', error: 'Server error.' });
 	}
 	
-	
-	let query = "UPDATE ALT_Database.express SET value=\'"+ req.body.value + "\' WHERE ID=" + req.params.id;
-	
-	console.log(query);
+	let query = "UPDATE ALT_Database.express SET value=\'" + req.body.value + "\' WHERE ID=" + req.params.id + " AND user_id=" + req.user.ID;
 	
 	connection.query(query, function(err, result) {
 		if (err) {
@@ -70,7 +71,7 @@ router.post('/:id', function(req, res, next) {
 
 router.delete('/:id', function(req,res, next) {
 	
-	let query = "DELETE FROM ALT_Database.express WHERE ID=" + req.params.id;
+	let query = "DELETE FROM ALT_Database.express WHERE ID=" + req.params.id + " AND user_id=" + req.user.ID;
 	
 	console.log(query);
